+++ docs/CHANGELOG.md
@@ -34,10 +34,12 @@
 
 ## 0 . 1 . 4
 
-+ 将 `LECTURE` 修改为 `HANDOUTS` 以更加符合语义
++ 将 `LECTURE` 修改为 `HANDOUTS`，更加符合语义
 
-+ 将 `explain` 方法名修改为 `solution` 以更加符合语义
++ 将 `explain` 方法名修改为 `solution`，更加符合语义
 
-+ `setup` 方法新增参数 `show-watermark` 添加水印功能
++ 修复当修改弥封线类型时,试卷最后一页没有更改的 `bug`
+
++ 添加水印功能，`setup` 方法新增参数 `show-watermark`，`watermark`，`watermark-size`，`watermark-color`，`watermark-font`，`watermark-rotate-degree`，`watermark-opacity`